+++ Sources/API/Sound/SoundProviders/soundfilter_provider.h
@@ -58,7 +58,7 @@ public:
 	    which modifies the sample data accordingly to the function of the
 	    filter.</p>
 	    <p>The format of the sample data is always 16 bit stereo. </p>*/
-	virtual void filter(int **sample_data, int num_samples, int channels)=0;
+	virtual void filter(float **sample_data, int num_samples, int channels)=0;
 
 /// \}
 /// \name Implementation
